Moved and cleaned TFModPlayer.
