+++ TerrariaFortress.cs
@@ -97,149 +97,6 @@ namespace TerrariaFortress
             SoulGargoyle
         }
 
-        public class TFModPlayer : ModPlayer
-        {
-            public float recentDamage = 0;
-            public float storedRandomCritChance = 0f;
-            /// <summary>
-            /// The max amount of damage stored for determining random crit chance.
-            /// </summary>
-            public float randomCritStorageCap = 2060f;
-            /// <summary>
-            /// If the player has crits active.
-            /// </summary>
-            public bool critBoosted = false;
-            /// <summary>
-            /// If the player has mini-crits active.
-            /// </summary>
-            public bool miniCritBoosted = false;
-            /// <summary>
-            /// If the player is rocket jumping.
-            /// </summary>
-            public bool isRocketJumping = false;
-            public float randomCritCheck = Main.rand.NextFloat();
-            public bool randomCrit = false;
-
-            public virtual bool ItemCritCheck(Player player, TFWeapon TFWeapon)
-            {
-                if (TFWeapon.canCrit)
-                {
-                    if (player.GetModPlayer<TFModPlayer>().randomCritCheck <= player.GetModPlayer<TFModPlayer>().storedRandomCritChance / 100f)
-                    {
-                        return true;
-                    }   
-
-                    if (player.GetModPlayer<TFModPlayer>().critBoosted)
-                    {
-                        return true;
-                    }
-                }
-                
-                return false;
-            }
-
-            public override void PreUpdate()
-            {
-                #region Random Crit Calculation
-                player.GetModPlayer<TFModPlayer>().randomCritCheck = Main.rand.NextFloat();
-                if (player.GetModPlayer<TFModPlayer>().recentDamage > 0f)
-                {
-                    player.GetModPlayer<TFModPlayer>().recentDamage -= 0.5f;
-                }
-                player.GetModPlayer<TFModPlayer>().storedRandomCritChance = MathHelper.Clamp((player.GetModPlayer<TFModPlayer>().recentDamage > 0 ? player.GetModPlayer<TFModPlayer>().recentDamage / 82.5f / 2f : 0) + 2, 2f, 12f);
-                player.GetModPlayer<TFModPlayer>().randomCrit = player.GetModPlayer<TFModPlayer>().randomCritCheck <= (storedRandomCritChance / 100f);
-                #endregion
-
-                #region Rocket Jumping
-                if (player.velocity.Y == 0f)
-                {
-                    if (player.GetModPlayer<TFModPlayer>().isRocketJumping)
-                    {
-                        player.velocity.X *= 0.8f;
-                        player.GetModPlayer<TFModPlayer>().isRocketJumping = false;
-                    }
-                }
-                #endregion
-            }
-
-            #region Random Crit Check
-
-            public override void ModifyHitNPC(Item item, NPC target, ref int damage, ref float knockback, ref bool crit)
-            {
-                if (target.CanBeChasedBy())
-                {
-                    if (item.modItem is TFWeapon TFWeapon)
-                    {
-                        crit = false;
-
-                        if (player.GetModPlayer<TFModPlayer>().recentDamage + damage < player.GetModPlayer<TFModPlayer>().randomCritStorageCap)
-                        {
-                            player.GetModPlayer<TFModPlayer>().recentDamage += damage;
-                        }
-                        else
-                        {
-                            player.GetModPlayer<TFModPlayer>().recentDamage += player.GetModPlayer<TFModPlayer>().randomCritStorageCap - player.GetModPlayer<TFModPlayer>().recentDamage;
-                        }
-                    }
-                }
-            }
-
-            public override void ModifyHitNPCWithProj(Projectile proj, NPC target, ref int damage, ref float knockback, ref bool crit, ref int hitDirection)
-            {
-                if (target.CanBeChasedBy())
-                {
-                    if (proj.modProjectile is TFProjectile TFProjectile)
-                    {
-                        crit = false;
-
-                        if (player.GetModPlayer<TFModPlayer>().recentDamage + damage < player.GetModPlayer<TFModPlayer>().randomCritStorageCap)
-                        {
-                            player.GetModPlayer<TFModPlayer>().recentDamage += damage;
-                        }
-                        else
-                        {
-                            player.GetModPlayer<TFModPlayer>().recentDamage += player.GetModPlayer<TFModPlayer>().randomCritStorageCap - player.GetModPlayer<TFModPlayer>().recentDamage;
-                        }
-                    }
-                }
-            }
-
-            public override void ModifyHitPvp(Item item, Player target, ref int damage, ref bool crit)
-            {
-                if (item.modItem is TFWeapon TFWeapon)
-                {
-                    crit = false;
-
-                    if (player.GetModPlayer<TFModPlayer>().recentDamage + damage < player.GetModPlayer<TFModPlayer>().randomCritStorageCap)
-                    {
-                        player.GetModPlayer<TFModPlayer>().recentDamage += damage;
-                    }
-                    else
-                    {
-                        player.GetModPlayer<TFModPlayer>().recentDamage += player.GetModPlayer<TFModPlayer>().randomCritStorageCap - player.GetModPlayer<TFModPlayer>().recentDamage;
-                    }
-                }
-            }
-
-            public override void ModifyHitPvpWithProj(Projectile proj, Player target, ref int damage, ref bool crit)
-            {
-                if (proj.modProjectile is TFProjectile TFProjectile)
-                {
-                    crit = false;
-
-                    if (player.GetModPlayer<TFModPlayer>().recentDamage + damage < player.GetModPlayer<TFModPlayer>().randomCritStorageCap)
-                    {
-                        player.GetModPlayer<TFModPlayer>().recentDamage += damage;
-                    }
-                    else
-                    {
-                        player.GetModPlayer<TFModPlayer>().recentDamage += player.GetModPlayer<TFModPlayer>().randomCritStorageCap - player.GetModPlayer<TFModPlayer>().recentDamage;
-                    }
-                }
-            }
-            #endregion
-        }
-
         public abstract class TFProjectile : ModProjectile
         {
             public bool willCrit = false;
